[Perf] Simplify ResourceRepositoryServiceTest.
Rewrite to mocks, remove tests duplicating DAO functionality.
